third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/disconnected-device.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => assert_promise_rejects_with_message(
      device.gatt.getPrimaryService('heart_rate'),
      new DOMException('GATT Server is disconnected. ' +
                         'Cannot retrieve services. ' +
                         '(Re)connect first with `device.gatt.connect`.',
                       'NetworkError')));
}, 'getPrimaryService() called before connecting. Reject with NetworkError.');
</script>
